pyproject.toml
[project]
name = "tasklistbot"
version = "0.2.0"
description = "Task list telegram bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiogram-dialog>=2.3.1",
    "aiogram-i18n[compiler]>=1.4",
    "beanie>=1.29.0",
    "cachetools>=5.5.2",
    "pydantic-settings>=2.9.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
